Added default roles.

+++ migrations/m180325_230019_custom_updates.php
@@ -12,10 +12,23 @@ class m180325_230019_custom_updates extends Migration
      */
     public function safeUp()
     {
-        $this->addColumn("user","firstname",$this->string(255)->notNull()->after("id"));
-        $this->addColumn("user","lastname",$this->string(255)->notNull()->after("firstname"));
-        $this->dropColumn("user", "username");
-        $this->dropTable("profile");
+        $this->addColumn("user", "firstname", $this->string(255)->notNull()->after("id"));
+        $this->addColumn("user", "lastname", $this->string(255)->notNull()->after("firstname"));
+
+        if (Yii::$app->db->schema->getTableSchema('user')->getColumn("username")){
+            $this->dropColumn("user", "username");
+        }
+
+        if (Yii::$app->db->schema->getTableSchema('profile')) {
+            $this->dropTable("profile");
+        }
+
+        /* ROLES */
+        $auth = Yii::$app->getAuthManager();
+        $superadmin = $auth->createRole('superadmin');
+        $auth->add($superadmin);
+        $admin = $auth->createRole('admin');
+        $auth->add($admin);
     }
 
     /**
@@ -23,8 +36,9 @@ class m180325_230019_custom_updates extends Migration
      */
     public function safeDown()
     {
-        $this->dropColumn("user","firstname");
-        $this->dropColumn("user","lastname");
+        $this->dropColumn("user", "firstname");
+        $this->dropColumn("user", "lastname");
+        $auth = Yii::$app->getAuthManager()->removeAll();
     }
 
 }
